Fixed logic after merge.
